Restructure and rename trending strategy files
