+++ discovery-provider/src/tasks/index_trending.py
@@ -5,8 +5,8 @@ from src.tasks.celery_app import celery
 from src.queries.get_trending_tracks import make_trending_cache_key, generate_unpopulated_trending
 from src.utils.redis_cache import pickle_and_set
 from src.utils.redis_constants import trending_tracks_last_completion_redis_key
-from src.utils.trending_selector import TrendingSelector
-from src.utils.trending_strategy import TrendingType
+from src.trending_strategies.trending_strategy_factory import TrendingStrategyFactory
+from src.trending_strategies.trending_type_and_version import TrendingType
 from src.queries.get_underground_trending import make_underground_trending_cache_key, make_get_unpopulated_tracks
 
 logger = logging.getLogger(__name__)
@@ -74,18 +74,18 @@ def get_genres(session):
     genres = filter(lambda x: x[0] is not None and x[0] != "" and x[0] in genre_allowlist, genres)
     return list(map(lambda x: x[0], genres))
 
+trending_strategy_factory = TrendingStrategyFactory()
 
 def index_trending(self, db, redis):
     logger.info('index_trending.py | starting indexing')
     update_start = time.time()
-    trending_selector = TrendingSelector()
     with db.scoped_session() as session:
         genres = get_genres(session)
 
         # Make sure to cache empty genre
         genres.append(None)
 
-        strategy = trending_selector.get_strategy(TrendingType.TRACKS)
+        strategy = trending_strategy_factory.get_strategy(TrendingType.TRACKS)
         for genre in genres:
             for time_range in time_ranges:
                 cache_start_time = time.time()
@@ -98,7 +98,7 @@ def index_trending(self, db, redis):
 
         # Cache underground trending
         cache_start_time = time.time()
-        strategy = trending_selector.get_strategy(TrendingType.UNDERGROUND_TRACKS)
+        strategy = trending_strategy_factory.get_strategy(TrendingType.UNDERGROUND_TRACKS)
         res = make_get_unpopulated_tracks(session, redis, strategy)()
         key = make_underground_trending_cache_key()
         pickle_and_set(redis, key, res)
