+++ discovery-provider/src/trending_strategies/trending_type_and_version.py
